HHH-20696 move test class to org.hibernate.orm.test.namingstrategy
